fix getmyplaylists because we could only submit one playlist at a time, handle case in deleteMyPlaylists, update css for delete playlist button on playlist page, add restricted and protected routes that user can and cant access when logged in, fix triggeranimation to trigger less often and remove once triggered or leave the page, separate loaders on user page for updating user info and deleting playlists, add discoverplaylist to display on home when no playlists submitted

+++ backend/controllers/userController.js
@@ -1,8 +1,6 @@
 const { User } = require('../model/User')
 const { Playlist } = require('../model/Playlist')
-
 const bcrypt = require('bcrypt');
-
 
 
 const updateUser = async (req, res) => {
@@ -39,18 +37,16 @@ const updateUser = async (req, res) => {
         if(!oldPassword && !newPassword) {
 
         } else {
-            console.log("\n res status set 2")
-
+            // this is handled on client side but just in case
             return res.status(404).json({"message": "Need both old and new password to update to new password"})
         }
     }
 
-
     const result = await userObject.save()
 
     res.status(200).json({
-         'newusername': result.userName,
-         'newemail': result.email
+         'newUsername': result.userName,
+         'newEmail': result.email
     })
 }
 
@@ -60,75 +56,31 @@ const deleteMyPlaylists = async (req, res) => {
     const userObject = await User.findOne({ _id: mongoUserId }).exec()
 
     var multipleIds = userObject.userPlaylistObjectIds.map((playlistObjectId) => { return { _id: playlistObjectId } })
-    const playlistObjects = await Playlist.find({ "$or":  multipleIds })
-
-
-    for (const playlistObject of playlistObjects) {
-        if(playlistObject.userOwner.length == 1) {
-            playlistObjectsToDelete.push(playlistObject._id.toString() ) // try without toString() afterwards
-        } else {
-            const newOwners = playlistObject.userOwner.filter((user) => user != mongoUserId)
-            playlistObject.userOwner = newOwners
-            const res3 = await playlistObject.save()
+    if (multipleIds.length) {
+        const playlistObjects = await Playlist.find({ "$or":  multipleIds })
+        for (const playlistObject of playlistObjects) {
+            if(playlistObject.userOwner.length == 1) {
+                playlistObjectsToDelete.push(playlistObject._id)
+                // try playlistObject._id.toString()
+            } else {
+                const newOwners = playlistObject.userOwner.filter((user) => user != mongoUserId)
+                playlistObject.userOwner = newOwners
+                const resPlaylistObject = await playlistObject.save()
+            }
         }
+
+        if(playlistObjectsToDelete) {
+            const resDeleteMany = await Playlist.deleteMany({ _id: { $in: playlistObjectsToDelete } });
+        }
+
+        userObject.userPlaylistObjectIds = []
+        const resSave = await userObject.save()
+    } else {
+        return res.status(204).json({ 'message': "nothing to delete" })
     }
-
-    if(playlistObjectsToDelete) {
-        console.log("\n\n\n\n these are the objects we will delete since only one owner")
-        console.log(playlistObjectsToDelete)
-
-        const resDeleteMany = await Playlist.deleteMany({ _id: { $in: playlistObjectsToDelete } }); 
-        console.log("\n\n\n resDeleteMany here")
-        console.log(resDeleteMany)
-    }
-
-    console.log("\n\n\n reached here 600")
-
-    // set user userPlaylistObjectIds to empty array here!!!!!!
-    // res cannot be sharing names with other stuff named res
-    userObject.userPlaylistObjectIds = []
-    const resSave = await userObject.save()
-
-    console.log("\n\n\n reached here 700")
-
 
     return res.status(200).json({ 'message': "success" })
 }
 
 
-
-
-
 module.exports = { updateUser, deleteMyPlaylists }
-
-
-
-
-
-    // playlistObjects.forEach((playlistObject) => {
-    //     // console.log("\n\n\n\n here is a single playlist objet")
-    //     // console.log(playlistObject.userOwner)
-    //     // console.log(playlistObject.userOwner.length)
-    //     // console.log(typeof(playlistObject.userOwner))})
-
-    //     if(playlistObject.userOwner.length == 1) {
-    //         console.log("\n\n\n\n\n this playlistObject only has one user")
-    //         console.log(playlistObject.userOwner)
-    //     } else {
-    //         console.log("\n\n\n\n this playlistObject has mulitple users")
-    //         console.log(playlistObject.userOwner)
-    //         const newOwners = playlistObject.userOwner.filter((user) => user != mongoUserId)
-
-    //         console.log("\n this newOwners")
-    //         console.log(newOwners)
-    //         playlistObject.userOwner = newOwners
-    //         await playlistObject.save()
-
-    //     }
-    // })
-
-
-    // const newPlaylistObjectIds = userObject.userPlaylistObjectIds.filter((object) => object.toString() !== playlistObjectId)
-
-
-
